refactor: moves code into individual modules
